backend/requirements.txt
# Core dependencies for Customer Complaint Triage Agent
anthropic>=0.18.0
python-dotenv>=1.0.0
requests>=2.31.0
jinja2>=3.1.3

# Email handling dependencies
imaplib2>=3.0.0
email-validator>=2.1.0

# Flask API dependencies
flask>=2.3.0
flask-cors>=4.0.0

# Optional dependencies for enhanced functionality
plotly>=5.15.0
pandas>=2.0.0

# Development and testing dependencies
pytest>=7.4.0
pytest-mock>=3.11.0
black>=23.0.0
flake8>=6.0.0
